Use id attributes for country and state anchor targets.
Update Issue 523
Status: Implemented
I used country and admin codes because to help ensure unique and valid IDs.

So United States gets id="US" and Wyoming gets id="US-WY". 

If there happen to be multiple lists on a page, a number will be appended to IDs after the first instance.


git-svn-id: https://wordpress-geo-mashup.googlecode.com/svn/branches/1.5-dev@1135 81465218-ea22-0410-8f2b-2343bde7f0e1