Sync geo_date with other geodata fields.
Update Issue 504
Status: Implemented
There's a tricky case to be aware of. If geo_date is saved after the other geodata custom fields, the object location will have already been added with the default date and the one added will not overwrite it. So if you have code to add geodata custom fields, add the coordinates last.


git-svn-id: https://wordpress-geo-mashup.googlecode.com/svn/branches/1.5-dev@1145 81465218-ea22-0410-8f2b-2343bde7f0e1