Fire endPan events when the map center changes.
Update Issue 598
Status: Fixed
Indeed, a little bug. Hopefully this will repair it in the next release.

I fear there will be more trouble with KML uploads in WordPress 3.5. I'll do my best to stay on top of it, but the many changes to the uploader may be challenging to keep up with.


git-svn-id: https://wordpress-geo-mashup.googlecode.com/svn/trunk@1151 81465218-ea22-0410-8f2b-2343bde7f0e1